create training.py
