+++ rasa/core/policies/rule_policy.py
@@ -39,7 +39,7 @@ from rasa.shared.core.constants import (
 from rasa.shared.core.domain import InvalidDomain, State, Domain
 from rasa.shared.nlu.constants import ACTION_NAME, INTENT_NAME_KEY
 import rasa.core.test
-import rasa.core.training
+import rasa.core.training.training
 
 
 if TYPE_CHECKING:
@@ -341,7 +341,7 @@ class RulePolicy(MemoizationPolicy):
     ) -> None:
         logger.debug("Started checking if some rules are incomplete.")
         # we need to use only fingerprints from rules
-        rule_fingerprints = rasa.core.training.create_action_fingerprints(
+        rule_fingerprints = rasa.core.training.training.create_action_fingerprints(
             rule_trackers, domain
         )
         if not rule_fingerprints:
